refactor: ProfileSetupModal 컴포넌트를 DropDownModal 합성 컴포넌트로 변경
