+++ src/features/users/components/DropDownModal/hooks/useDropDownModal.ts
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+export default function useDropDownModal() {
+  const [isDropDownModalOpen, setIsDropDownModalOpen] = useState(false);
+  const handleDropDownModalToggle = () =>
+    setIsDropDownModalOpen((prev) => !prev);
+
+  return { isDropDownModalOpen, handleDropDownModalToggle };
+}
